+++ test/dummy/admin/app/controllers/admin/authorized_articles_controller.rb
@@ -8,6 +8,20 @@ module Admin
       "admin"
     end
 
+    def new
+      # The following calls to #policy are to check that the Authorization
+      # module can handle various different scenarios:
+      policy(Magazine).index?
+      policy(::Magazine).index?
+      policy(Magazine.new).index?
+      policy(Magazine.all).index?
+      policy(Admin::Magazine).index?
+      policy(Admin::Magazine.new).index?
+      policy(Admin::Magazine.where(name: "name")).index?
+
+      super
+    end
+
     def resource_service_class
       Admin::ArticleService
     end