+++ test/dummy/admin/app/models/admin/magazine.rb
@@ -1,0 +1,4 @@
+module Admin
+  class Magazine < ::Magazine
+  end
+end