+++ test/dummy/app/models/magazine.rb
@@ -1,0 +1,2 @@
+class Magazine < ActiveRecord::Base
+end